+++ .eslintrc.js
@@ -9,11 +9,15 @@ module.exports = {
     '@vue/typescript/recommended',
   ],
   parserOptions: {
-    ecmaVersion: 2020,
+    "parser": "@typescript-eslint/parser"
   },
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'import/prefer-default-export': 'off',
+    semi: 'warn',
+    'comma-dangle': 'warn',
+    'no-trailing-spaces': 'off'
   },
   overrides: [
     {