+++ src/specification/api/GoodGroup.ts
@@ -1,0 +1,3 @@
+import { routes } from '@/specification/api/GoodRoutes';
+
+export type GoodGroup = keyof typeof routes;