+++ src/specification/api/GoodRoutes.ts
@@ -1,0 +1,5 @@
+export const routes = {
+  tShirts: 't-shirts',
+  trinkets: 'trinkets',
+  books: 'books',
+} as const;